- setup now runs PHP check script and opens it in a browser
git-svn-id: file:///home/smlewis/svn/git-mp3/another_svn/tmp/tmp/trunk@233 5035da28-9fce-0310-abbb-aec50abfc0e8

+++ molprobity3/setup.sh
@@ -13,7 +13,7 @@ if [ ! -f setup.sh ]; then
         echo
         echo "    You must run SETUP.SH from the main MolProbity directory!"
         echo
-        exit
+        exit 1
     fi
 fi
 
@@ -46,9 +46,29 @@ chmod 666 feedback/user_paths.log
 cd bin
 # Nothing to do here...
 cd macosx/
-ln -s gawk awk
+[ -f awk ] || ln -s gawk awk
 cd ..
 cd linux/
 # Nothing to do here...
 cd ..
 cd ..
+
+# Try to find PHP
+if ! type -p php > /dev/null; then
+    echo
+    echo "    Can't find PHP.  You must have PHP on your PATH to run MolProbity."
+    echo
+    exit 1
+fi
+php -C -f public_html/admin/check_config.php > tmp/check_config.html
+
+# Run setup check and show results
+if type -p open > /dev/null; then
+    open tmp/check_config.html
+elif type -p firefox > /dev/null; then
+    firefox tmp/check_config.html
+elif type -p mozilla > /dev/null; then
+    mozilla tmp/check_config.html
+else
+    echo "Please open ./tmp/check_config.html in a web browser."
+fi
